Move Graql pattern classes into pattern package
